requirements.txt
gdown==5.1.0
icecream==2.1.3
imblearn==0.0
numpy==1.24.4
pandas==2.0.3
PyGithub==2.1.1
PyYAML==5.3.1
scikit_learn==1.2.2
scipy==1.10.1
setuptools==67.8.0
torch==2.1.0
GitPython==3.1.7
PyDriller==1.15.2
options==1.4.10
testresources==2.0.1
dateparser==0.7.6
networkx==2.6.3
beautifulsoup4==4.10.0
lxml==4.6.4
packaging==21.3
regex==2022.1.18
python-Levenshtein==0.12.2
matplotlib==3.6.3
tqdm==4.67.1